refactor: 폴더 구조 변경
```
- 각자 이름 하위의 폴더에서 작업하도록 변경
```
